[Refactor] Move H3 native LoRA loader into a minimax_h3/npu package
Review follow-up. The Turbo and native loaders sat in one lora.py, but they
share no code: they differ in rank, target naming, QKV layout and metadata
contract, so keeping them together only made either one harder to read.

Move the native loader to vllm_omni/diffusion/models/minimax_h3/npu/lora.py, a
package grouping components for artifacts produced on Ascend NPU, and leave the
LightX2V Turbo loader in minimax_h3/lora.py. This is a pure relocation: no
symbol is renamed and no logic changes, so the existing suite pins it. Baseline
and post-move runs of tests/diffusion/models/minimax_h3/ report an identical
failure set, differing only by the tests added below.

Also answer whether the loader is NPU-only, since the package name could be
read as a runtime restriction. It is not: nothing here imports torch_npu, the
whole suite runs in the CPU job, and a dtype-parametrized test pins that every
tensor lands on CPU in exactly the requested dtype, leaving placement to the
pipeline. A CUDA-gated test recomputes all 415 lora_b @ lora_a products,
including the reordered Q/K/V slices, on GPU and compares them against the CPU
result so a GPU deployment cannot silently diverge.
